% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addImage.R
\name{add4ChannelImage}
\alias{add4ChannelImage}
\title{Add image to SpatialOverlay from disk}
\usage{
add4ChannelImage(overlay, ometiff = NULL, res = NULL, ...)
}
\arguments{
\item{overlay}{SpatialOverlay object}

\item{res}{resolution layer, 1 = largest & higher values = smaller. The images increase 
in resolution and memory. The largest image your environment 
can hold is recommended. NULL indicates pull info from overlay}

\item{imageFile}{path to image. NULL indicates pull infor from overlay}
}
\value{
SpatialOverlay object with image
}
\description{
Add image to SpatialOverlay from disk
}
\examples{

}
