% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addImage.R
\name{addImageFile}
\alias{addImageFile}
\title{Add image to SpatialOverlay from disk}
\usage{
addImageFile(overlay, imageFile, res, ...)
}
\arguments{
\item{overlay}{SpatialOverlay object}

\item{imageFile}{path to image}

\item{res}{resolution layer, 1 = largest & higher values = smaller. The images increase 
in resolution and memory. The largest image your environment 
can hold is recommended.}
}
\value{
SpatialOverlay object with image
}
\description{
Add image to SpatialOverlay from disk
}
\examples{

}
