% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPlottingFactor.R
\name{addPlottingFactor,data.frame-method}
\alias{addPlottingFactor,data.frame-method}
\title{Add plotting factor to \code{\link{SpatialOverlay}} object}
\usage{
\S4method{addPlottingFactor}{data.frame}(overlay, annots, plottingFactor)
}
\arguments{
\item{overlay}{\code{\link{SpatialOverlay}} object}

\item{annots}{\code{data.frame} containing plottingFactor, column or column names 
must contian matching sample names to overlay}

\item{plottingFactor}{column or row name from annots to add as plotting factor}
}
\value{
\code{\link{SpatialOverlay}} object with new plotting factor
}
\description{
Add plotting factor to \code{\link{SpatialOverlay}} object
}
\examples{

muBrain <- readRDS(unzip(system.file("extdata", "muBrain_SpatialOverlay.zip", 
                                    package = "SpatialOmicsOverlay")))

muBrainLW <- system.file("extdata", "muBrain_LabWorksheet.txt", 
                         package = "SpatialOmicsOverlay")

muBrainLW <- readLabWorksheet(muBrainLW, slideName = "4")

muBrain <- addPlottingFactor(overlay = muBrain, 
                             annots = muBrainLW, 
                             plottingFactor = "segment")


}
