% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPlottingFactor.R
\name{addPlottingFactor,factor-method}
\alias{addPlottingFactor,factor-method}
\title{Add plotting factor to \code{\link{SpatialOverlay}} object}
\usage{
\S4method{addPlottingFactor}{factor}(overlay, annots, plottingFactor)
}
\arguments{
\item{overlay}{\code{\link{SpatialOverlay}} object}

\item{annots}{factor vector with the plottingFactor. if names match sample 
names in overlay vector will be matched on those, 
otherwise assumed in the correct order}

\item{plottingFactor}{name of the new plotting factor}
}
\value{
\code{\link{SpatialOverlay}} object with new plotting factor
}
\description{
Add plotting factor to \code{\link{SpatialOverlay}} object
}
\examples{

muBrain <- readRDS(unzip(system.file("extdata", "muBrain_SpatialOverlay.zip", 
                                    package = "SpatialOmicsOverlay")))

muBrain <- addPlottingFactor(overlay = muBrain,
                             annots = as.factor(1:length(sampNames(muBrain))),
                             plottingFactor = "ROINum")

head(plotFactors(muBrain))

}
