% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPlottingFactor.R
\name{addPlottingFactor,NanoStringGeoMxSet-method}
\alias{addPlottingFactor,NanoStringGeoMxSet-method}
\title{Add plotting factor to \code{\link{SpatialOverlay}} object}
\usage{
\S4method{addPlottingFactor}{NanoStringGeoMxSet}(overlay, annots, plottingFactor, countMatrix = "exprs")
}
\arguments{
\item{overlay}{\code{\link{SpatialOverlay}} object}

\item{annots}{\code{NanoStringGeoMxSet} object}

\item{plottingFactor}{column or row from annots to add as plotting factor}

\item{countMatrix}{name of count matrix to pull counts from}
}
\value{
\code{\link{SpatialOverlay}} object with new plotting factor
}
\description{
Add plotting factor to \code{\link{SpatialOverlay}} object
}
\examples{

muBrain <- readRDS(unzip(system.file("extdata", "muBrain_SpatialOverlay.zip", 
                                    package = "SpatialOmicsOverlay")))

muBrainGxT <- readRDS(system.file("extdata", "muBrain_GxT.RDS", 
                                  package = "SpatialOmicsOverlay"))

muBrain <- addPlottingFactor(overlay = muBrain, 
                             annots = muBrainGxT, 
                             plottingFactor = "Calm1",
                             countMatrix = "exprs")

head(plotFactors(muBrain))

}
