% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBioFormatsFunctions.R
\docType{class}
\name{AnnotatedImage-class}
\alias{AnnotatedImage-class}
\alias{AnnotatedImage}
\alias{print.AnnotatedImage}
\alias{show,AnnotatedImage-method}
\alias{as.Image.AnnotatedImage}
\title{AnnotatedImage Class, from RBioFormats package}
\usage{
AnnotatedImage(...)

\method{print}{AnnotatedImage}(x, short = FALSE, ...)

\S4method{show}{AnnotatedImage}(object)

\method{as.Image}{AnnotatedImage}(x)
}
\arguments{
\item{...}{data to include in the new object. Named arguments correspond to 
slots in the class definition. Unnamed arguments must be objects 
from classes that this class extends.}

\item{x}{an \linkS4class{AnnotatedImage} object.}

\item{short}{logical, turns off image data preview.}

\item{object}{an \code{AnnotatedImage} object}
}
\value{
Extends the \linkS4class{Image} class from the \pkg{EBImage} package.
}
\description{
AnnotatedImage Class, from RBioFormats package
}
\section{Slots}{

\describe{
\item{\code{metadata}}{an \linkS4class{ImageMetadata} object containing image metadata}
}}

\references{
RBioFormats https://github.com/aoles/RBioFormats
}
\seealso{
\linkS4class{AnnotatedImageList}
}
