% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBioFormatsFunctions.R
\docType{class}
\name{AnnotatedImageList-class}
\alias{AnnotatedImageList-class}
\alias{AnnotatedImageList}
\title{AnnotatedImageList Class}
\usage{
AnnotatedImageList(...)
}
\arguments{
\item{...}{a list of \linkS4class{AnnotatedImage} objects to include in the new object.}
}
\value{
A list of \linkS4class{AnnotatedImage} objects.
}
\description{
AnnotatedImageList Class
}
\references{
RBioFormats https://github.com/aoles/RBioFormats
}
