% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageManipulation.R
\name{autoCrop}
\alias{autoCrop}
\title{Create coordinate file for entire scan}
\usage{
autoCrop(object)
}
\arguments{
\item{overlay}{df from GeoMxmlParsing_AOIattrs}

\item{outline}{returned coordinates only contain outlinearies, 
will not work for segmented ROIs}
}
\value{
df of coordinates for every AOI in the scan
}
\description{
Create coordinate file for entire scan
}
\examples{

}
