% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omeExtraction.R
\name{checkValidRes}
\alias{checkValidRes}
\title{Determine lowest resolution image in OME-TIFF}
\usage{
checkValidRes(ometiff)
}
\arguments{
\item{ometiff}{path to OME-TIFF}
}
\value{
value of lowest res image
}
\description{
Determine lowest resolution image in OME-TIFF
}
\examples{

image <- downloadMouseBrainImage()
checkValidRes(ometiff = image)

}
