% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinateGeneration.R
\name{coordsFromMask}
\alias{coordsFromMask}
\title{Return positive coordinates from mask}
\usage{
coordsFromMask(mask, metadata, outline = TRUE)
}
\arguments{
\item{mask}{base 64 string}

\item{metadata}{metadata of AOI including: X,Y position in full image}

\item{outline}{only the outline points should be returned}
}
\value{
df of positive coordinates of AOI
}
\description{
Return positive coordinates from mask
}
\examples{

}
