% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinateGeneration.R
\name{createCoordFile}
\alias{createCoordFile}
\title{Create coordinate file for entire scan}
\usage{
createCoordFile(overlay, outline = TRUE)
}
\arguments{
\item{overlay}{SpatialOverlay object}

\item{outline}{returned coordinates only contain boundaries, 
will not work for segmented ROIs}
}
\value{
df of coordinates for every AOI in the scan
}
\description{
Create coordinate file for entire scan
}
\examples{

muBrain <- readRDS(system.file("extdata", "muBrain_SpatialOverlay.RDS", 
                               package = "SpatialOmicsOverlay"))

muBrain <- createCoordFile(muBrain, outline = FALSE)

head(coords(muBrain))

}
