% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlParsing.R
\name{decodeB64}
\alias{decodeB64}
\title{Decode a base 64 string into 8 bit binary image mask}
\usage{
decodeB64(b64string, width, height)
}
\arguments{
\item{b64string}{base 64 string}

\item{width}{width of image}

\item{height}{height of image}
}
\value{
binary mask image
}
\description{
Decode a base 64 string into 8 bit binary image mask
}
\examples{

}
