% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downloadMouseBrainImage}
\alias{downloadMouseBrainImage}
\title{Download Mouse Brain OME-TIFF from NanoString's Spatial Organ Atlas}
\usage{
downloadMouseBrainImage()
}
\value{
mouse brain OME-TIFF
}
\description{
Download Mouse Brain OME-TIFF from NanoString's Spatial Organ Atlas
}
\details{
https://nanostring.com/products/geomx-digital-spatial-profiler/spatial-organ-atlas/mouse-brain/
}
\examples{

image <- downloadMouseBrainImage()

}
