% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageManipulation.R
\name{flipX}
\alias{flipX}
\title{Flip x axis in image and overlay points}
\usage{
flipX(overlay)
}
\arguments{
\item{overlay}{SpatialOverlay object}
}
\value{
SpatialOverlay object with x axis flipped
}
\description{
Flip x axis in image and overlay points
}
\examples{

muBrain <- readRDS(unzip(system.file("extdata", "muBrainSubset_SpatialOverlay.zip", 
                                    package = "SpatialOmicsOverlay")))

image <- downloadMouseBrainImage()

muBrain <- addImageOmeTiff(overlay = muBrain, 
                           ometiff = image, res = 8)

showImage(flipX(muBrain))

}
