% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlParsing.R
\name{fluorData}
\alias{fluorData}
\title{Parse fluorescence data from xml}
\usage{
fluorData(omexml)
}
\arguments{
\item{omexml}{xml file from OME-TIFF}
}
\value{
fluorescence info from xml
}
\description{
Parse fluorescence data from xml
}
\examples{

image <- downloadMouseBrainImage()

xml <- xmlExtraction(ometiff = image)

fluorData(xml)

}
