% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageManipulation.R
\name{imageColoring}
\alias{imageColoring}
\title{Turn 4-channel composite image into RGB}
\usage{
imageColoring(omeImage, scanMeta)
}
\arguments{
\item{omeImage}{AnnotatedImage from read.image}

\item{scanMeta}{scanMeta data from SpatialOverlay}
}
\value{
AnnotatedImage with RGB matrix
}
\description{
Turn 4-channel composite image into RGB
}
