% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omeExtraction.R
\name{imageExtraction}
\alias{imageExtraction}
\title{Extract image from OME-TIFF}
\usage{
imageExtraction(
  ometiff,
  res = 6,
  scanMeta = NULL,
  saveFile = FALSE,
  fileType = "tiff",
  color = TRUE,
  outdir = NULL
)
}
\arguments{
\item{ometiff}{path to OME-TIFF}

\item{res}{resolution layer, 1 = largest & higher values = smaller. The 
images increase in resolution and memory. The largest image 
your environment can hold is recommended.}

\item{scanMeta}{scan metadata from parseScanMetadata(). If NULL, that 
function is automatically called.}

\item{saveFile}{should xml be saved, file is saved in working directory with 
same name as OME-TIFF}

\item{fileType}{type of image file. Options: tiff, png, jpeg}

\item{color}{should image be colored RGB or 4-channel BW}

\item{outdir}{output directory for saved image. If NULL, saved in same directory as OME-TIFF}
}
\value{
omeImage magick pointer
}
\description{
Extract image from OME-TIFF
}
\examples{

image <- downloadMouseBrainImage()

imageExt <- imageExtraction(ometiff = image, res = 7)
imageExt

}
