% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBioFormatsFunctions.R
\docType{class}
\name{ImageMetadata-class}
\alias{ImageMetadata-class}
\alias{ImageMetadataList-class}
\alias{ImageMetadata}
\alias{ImageMetadataList}
\alias{print.ImageMetadata}
\alias{show,ImageMetadata-method}
\alias{show,ImageMetadataList-method}
\alias{print.ImageMetadataList}
\title{ImageMetadata and ImageMetadataList Class, from RBioFormats package}
\usage{
ImageMetadata(...)

ImageMetadataList(...)

\method{print}{ImageMetadata}(x, list.len = 5L, ...)

\S4method{show}{ImageMetadata}(object)

\S4method{show}{ImageMetadataList}(object)

\method{print}{ImageMetadataList}(x, ...)
}
\arguments{
\item{...}{further arguments to be passed to other methods}

\item{x}{An ImageMetadata object}

\item{list.len}{numeric; maximum number of metadata entries to display}

\item{object}{an \code{ImageMetadata} object}
}
\value{
Formal representation of image metadata.
}
\description{
ImageMetadata and ImageMetadataList Class, from RBioFormats package
}
\references{
RBioFormats https://github.com/aoles/RBioFormats
}
