% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlParsing.R
\name{parseScanMetatdata}
\alias{parseScanMetatdata}
\title{Parse the xml file for the scan metadata of GeoMx images}
\usage{
parseScanMetatdata(omexml)
}
\arguments{
\item{omexml}{xml file from OME-TIFF, can provide path to OME-TIFF and 
xml will automatically be extracted}
}
\value{
metadata for entire scan
}
\description{
Parse the xml file for the scan metadata of GeoMx images
}
\examples{

image <- downloadMouseBrainImage()

xml <- xmlExtraction(ometiff = image)

scan_metadata <- parseScanMetatdata(omexml = xml)

}
