% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinateGeneration.R
\name{pencilCoordSorting}
\alias{pencilCoordSorting}
\title{Sort coordinates like a pencil drawing}
\usage{
pencilCoordSorting(coords, rangeWidth = 100)
}
\arguments{
\item{coords}{All x,y coordinates for an ROI}

\item{rangeWidth}{size of range to search for closest coordinate}
}
\value{
df of ordered coordinates
}
\description{
When outline coordinates are generated they are ordered by column but 
geom_polygon expects them to be in order like you are tracing the shape 
with a pencil. This function orders them by finding the closest x,y 
coordinate to the previous point
}
\examples{

}
