% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlParsing.R
\name{physicalSizes}
\alias{physicalSizes}
\title{Parse physicalSize data from xml}
\usage{
physicalSizes(omexml)
}
\arguments{
\item{omexml}{xml file from OME-TIFF}
}
\value{
physicalSizes info from xml
}
\description{
Parse physicalSize data from xml
}
\examples{

image <- downloadMouseBrainImage()

xml <- xmlExtraction(ometiff = image)

physicalSizes(xml)

}
