% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readLabWorksheet}
\alias{readLabWorksheet}
\title{Read lab worksheet into dataframe of annotations}
\usage{
readLabWorksheet(lw, slideName)
}
\arguments{
\item{lw}{lab worksheet file path}

\item{slideName}{name of slide}
}
\value{
df of ROI annotations
}
\description{
Read lab worksheet into dataframe of annotations
}
\examples{

muBrainLW <- system.file("extdata", "muBrain_LabWorksheet.txt", 
                         package = "SpatialOmicsOverlay")

muBrainLW <- readLabWorksheet(muBrainLW, slideName = "4")

}
