% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageManipulation.R
\name{recolor}
\alias{recolor}
\title{recolor images after changing colors and/or color intensities}
\usage{
recolor(overlay)
}
\arguments{
\item{overlay}{SpatialOverlay object}
}
\value{
SpatialOverlay object with RGB image
}
\description{
recolor images after changing colors and/or color intensities
}
\examples{

}
