% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSamples.R
\name{removeSample}
\alias{removeSample}
\title{Remove sample(s) from SpatialOverlay}
\usage{
removeSample(object, remove)
}
\arguments{
\item{object}{SpatialOverlay object}

\item{remove}{sampNames of overlay to remove}
}
\value{
SpatialOverlay object without samples in remove
}
\description{
Remove sample(s) from SpatialOverlay
}
\examples{

}
