% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingBasics.R
\name{scaleBarPrinting}
\alias{scaleBarPrinting}
\title{Add scale bar to ggplot object}
\usage{
scaleBarPrinting(
  gp,
  scaleBar,
  corner = "bottomright",
  scaleBarFontSize = 6,
  scaleBarLineSize = 1.5,
  scaleBarColor = "red",
  textDistance = 2,
  image = NULL,
  ...
)
}
\arguments{
\item{gp}{ggplot object}

\item{scaleBar}{output from scaleBarMath}

\item{corner}{where in the figure should the scale bar be printed. 
Options: "bottomright"  "topright"
         "bottomleft"   "topleft"
         "bottomcenter" "topcenter"}

\item{scaleBarFontSize}{scale bar font size}

\item{scaleBarLineSize}{scale bar width of line}

\item{scaleBarColor}{scale bar color}

\item{textDistance}{text's distance from scale bar.}

\item{image}{image from SpatialOverlay or NULL}

\item{...}{additional parameters for geom_line & geom_text, will apply to both}
}
\value{
gp with scale bar
}
\description{
must run scaleBarMath() first
}
