% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinateGeneration.R
\name{scaleCoords}
\alias{scaleCoords}
\title{Scale coordinates to the size of the image}
\usage{
scaleCoords(overlay)
}
\arguments{
\item{overlay}{SpatialOverlay object}
}
\value{
SpatialOverlay object
}
\description{
Scale coordinates to the size of the image
}
\examples{

muBrain <- readRDS(unzip(system.file("extdata", "muBrain_SpatialOverlay.zip", 
                                    package = "SpatialOmicsOverlay")))

image <- downloadMouseBrainImage()

muBrain <- addImageOmeTiff(overlay = muBrain,
                           ometiff = image, res = 7)

summary(coords(muBrain))

muBrain <- scaleCoords(overlay = muBrain)

#fewer and smaller points
summary(coords(muBrain))

}
