% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingBasics.R
\name{themeTransparent}
\alias{themeTransparent}
\title{transparent theme for plots}
\usage{
themeTransparent()
}
\value{
ggplot theme
}
\description{
transparent theme for plots
}
\examples{

muBrain <- readRDS(unzip(system.file("extdata", "muBrain_SpatialOverlay.zip", 
                                    package = "SpatialOmicsOverlay")))

muBrainLW <- system.file("extdata", "muBrain_LabWorksheet.txt", 
                         package = "SpatialOmicsOverlay")

muBrainLW <- readLabWorksheet(muBrainLW, slideName = "4")

image <- downloadMouseBrainImage()

muBrain <- addImageOmeTiff(overlay = muBrain, 
                          ometiff = image, res = 7)

muBrain <- addPlottingFactor(overlay = muBrain, 
                             annots = muBrainLW, 
                             plottingFactor = "segment")
                             
plotSpatialOverlay(overlay = muBrain, colorBy = "segment",  
                   hiRes = TRUE, scaleBar = FALSE) + 
themeTransparent()

}
