% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addImage.R
\name{add4ChannelImage}
\alias{add4ChannelImage}
\title{Add 4-channel image to SpatialOverlay from OME-TIFF. Allows for recoloring 
of image}
\usage{
add4ChannelImage(overlay, ometiff = NULL, res = NULL, ...)
}
\arguments{
\item{overlay}{SpatialOverlay object}

\item{ometiff}{File path to OME-TIFF. NULL indicates pull info from overlay}

\item{res}{resolution layer, 1 = largest & higher values = smaller. The 
images increase in resolution and memory. The largest image your 
environment can hold is recommended.  NULL indicates pull info 
from overlay}

\item{...}{Extra variables for \code{\link{imageExtraction}}}
}
\value{
SpatialOverlay object with image
}
\description{
Add 4-channel image to SpatialOverlay from OME-TIFF. Allows for recoloring 
of image
}
\examples{

muBrain <- readRDS(unzip(system.file("extdata", "muBrain_SpatialOverlay.zip", 
                                    package = "SpatialOmicsOverlay")))
image <- downloadMouseBrainImage()

muBrain <- add4ChannelImage(overlay = muBrain, 
                            ometiff = image, res = 7)

dim(EBImage::imageData(showImage(muBrain)))

}
