% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addImage.R
\name{addImageFile}
\alias{addImageFile}
\title{Add image to SpatialOverlay from disk}
\usage{
addImageFile(overlay, imageFile = NULL, res = NULL)
}
\arguments{
\item{overlay}{SpatialOverlay object}

\item{imageFile}{path to image}

\item{res}{what resolution is the image given? 
1 = largest, higher number = smaller
This value will affect the coordinates of the overlays. 
res = 2, resolution is 1/2 the size as the raw image
res = 3, resolution is 1/4 the size as the raw image
res = 4, resolution is 1/8 the size as the raw image 
resolution = 1/2^(res-1)}
}
\value{
SpatialOverlay object with image
}
\description{
Add image to SpatialOverlay from disk
}
