% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPlottingFactor.R
\name{addPlottingFactor}
\alias{addPlottingFactor}
\title{Add plotting factor to \code{\link{SpatialOverlay}} object}
\usage{
addPlottingFactor(overlay, annots, plottingFactor, ...)
}
\arguments{
\item{overlay}{\code{\link{SpatialOverlay}} object}

\item{annots}{object with annotation: can be \code{NanoStringGeoMxSet}, 
\code{data.frame}, \code{matrix}, or vector containing 
\code{character}, \code{factor}, or \code{numeric}}

\item{plottingFactor}{column or row from annots to add as plotting factor, 
if using vector this will be the name of the new 
plotting factor}

\item{...}{if using \code{NanoStringGeoMxSet}, name of count matrix to pull 
counts from}
}
\value{
\code{\link{SpatialOverlay}} object with new plotting factor
}
\description{
Add plotting factor to \code{\link{SpatialOverlay}} object
}
\examples{

muBrain <- readRDS(unzip(system.file("extdata", "muBrain_SpatialOverlay.zip", 
                                    package = "SpatialOmicsOverlay")))

muBrainLW <- system.file("extdata", "muBrain_LabWorksheet.txt", 
                         package = "SpatialOmicsOverlay")

muBrainLW <- readLabWorksheet(muBrainLW, slideName = "4")

muBrain <- addPlottingFactor(overlay = muBrain, 
                             annots = muBrainLW, 
                             plottingFactor = "segment")

library(GeomxTools)
muBrainGxT <- readRDS(system.file("extdata", "muBrain_GxT.RDS", 
                                  package = "SpatialOmicsOverlay"))

muBrain <- addPlottingFactor(overlay = muBrain, 
                             annots = muBrainGxT, 
                             plottingFactor = "Calm1",
                             countMatrix = "exprs")

muBrain <- addPlottingFactor(overlay = muBrain,
                             annots = seq_len(length(sampNames(muBrain))),
                             plottingFactor = "ROINum")

head(plotFactors(muBrain))

}
