% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageManipulation.R
\name{changeImageColoring}
\alias{changeImageColoring}
\title{Update color scheme for changing to RGB image}
\usage{
changeImageColoring(overlay, color, dye)
}
\arguments{
\item{overlay}{SpatialOverlay object, with 4channel image}

\item{color}{color to change dye to, hex or color name}

\item{dye}{which dye to change color, can be from Dye or DisplayName column from fluor(overlay)}
}
\value{
SpatialOverlay object with updated fluor data
}
\description{
Update color scheme for changing to RGB image
}
\examples{

}
