% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageManipulation.R
\name{crop}
\alias{crop}
\title{Create coordinate file for entire scan}
\usage{
crop(overlay, xmin, xmax, ymin, ymax, coords = TRUE)
}
\arguments{
\item{overlay}{SpatialOverlay object}

\item{xmin}{minimum x coordinate}

\item{xmax}{maximum x cooridnate}

\item{ymin}{minimum y coordinate}

\item{ymax}{maximum y coordinate}

\item{coords}{should coords be cropped}
}
\value{
df of coordinates for every AOI in the scan
}
\description{
Create coordinate file for entire scan
}
\examples{

}
