% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageManipulation.R
\name{cropTissue}
\alias{cropTissue}
\title{Crop to remove black boundary around tissue.}
\usage{
cropTissue(overlay, buffer = 0.05)
}
\arguments{
\item{overlay}{SpatialOverlay object}

\item{buffer}{percent of new image size to add to each edge as a buffer}
}
\value{
SpatialOverlay object
}
\description{
Crop to remove black boundary around tissue.
}
\examples{
#' muBrain <- readRDS(system.file("extdata", "muBrain_SpatialOverlay.RDS", 
                               package = "SpatialOmicsOverlay"))

muBrainLW <- system.file("extdata", "muBrain_LabWorksheet.txt", 
                         package = "SpatialOmicsOverlay")

muBrainLW <- readLabWorksheet(muBrainLW, slideName = "4")

image <- downloadMouseBrainImage()

muBrain <- addImageOmeTiff(overlay = muBrain, 
                           ometiff = image, res = 7)

muBrain <- addPlottingFactor(overlay = muBrain, 
                             annots = muBrainLW, 
                             plottingFactor = "segment")

samps <- muBrainLW$Sample_ID[muBrainLW$segment == "Full ROI"]

muBrainCrop <- cropTissue(overlay = muBrain)

plotSpatialOverlay(overlay = muBrainCrop, colorBy = "segment",  
                   hiRes = TRUE, scaleBar = F)
                   
}
