% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageManipulation.R
\name{flipX}
\alias{flipX}
\title{Flip x axis in image and overlay points}
\usage{
flipX(overlay)
}
\arguments{
\item{overlay}{SpatialOverlay object}
}
\value{
SpatialOverlay object with x axis flipped
}
\description{
Flip x axis in image and overlay points
}
\examples{

muBrain <- readRDS(system.file("extdata", "muBrainSubset_SpatialOverlay.RDS", 
                                    package = "SpatialOmicsOverlay"))

muBrainLW <- system.file("extdata", "muBrain_LabWorksheet.txt", 
                         package = "SpatialOmicsOverlay")

muBrainLW <- readLabWorksheet(muBrainLW, slideName = "4")

image <- downloadMouseBrainImage()

muBrain <- addImageOmeTiff(overlay = muBrain, 
                           ometiff = image, res = 8)

muBrain <- addPlottingFactor(overlay = muBrain, 
                             annots = muBrainLW, 
                             plottingFactor = "segment")

plotSpatialOverlay(overlay = flipX(muBrain), colorBy = "segment",  
                   hiRes = TRUE, scaleBar = FALSE)

}
