% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSpatialOverlay.R
\name{readSpatialOverlay}
\alias{readSpatialOverlay}
\title{Read in \code{\linkS4class{SpatialOverlay}} from tiff file and annotations}
\usage{
readSpatialOverlay(
  ometiff,
  annots,
  slideName,
  image = FALSE,
  res = NULL,
  saveFile = FALSE,
  outline = TRUE
)
}
\arguments{
\item{ometiff}{path to OME-TIFF}

\item{annots}{path to annotation file: can be labWorksheet, DA excel file, or delimted file}

\item{slideName}{name of slide}

\item{image}{should image be extracted from OME-TIFF}

\item{res}{resolution of image \code{\link{imageExtraction}}}

\item{saveFile}{should xml & image be saved, file is saved in working directory 
with same name as OME-TIFF}

\item{outline}{returned coordinates only contain outlinearies, 
will not work for segmented ROIs}
}
\value{
\code{\linkS4class{SpatialOverlay}} of slide
}
\description{
Create an instance of class \code{\linkS4class{SpatialOverlay}} 
by reading data from OME-TIFF and annotation sheet.
}
\examples{

muBrainLW <- system.file("extdata", "muBrain_LabWorksheet.txt", 
                         package = "SpatialOmicsOverlay")

image <- downloadMouseBrainImage()

muBrain <- readSpatialOverlay(ometiff = image, annots = muBrainLW, 
                              slideName = "4", image = TRUE, res = 7, 
                              saveFile = FALSE, outline = FALSE)

}
\seealso{
\code{\link{SpatialOverlay-class}}
}
