% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageManipulation.R
\name{recolor}
\alias{recolor}
\title{recolor images after changing colors and/or color intensities}
\usage{
recolor(overlay)
}
\arguments{
\item{overlay}{SpatialOverlay object}
}
\value{
SpatialOverlay object with RGB image
}
\description{
recolor images after changing colors and/or color intensities
}
\examples{

muBrain <- readRDS(unzip(system.file("extdata", "muBrainSubset_SpatialOverlay.zip", 
                                    package = "SpatialOmicsOverlay")))

image <- downloadMouseBrainImage()

muBrain <- add4ChannelImage(overlay = muBrain, 
                            ometiff = image, res = 8)

muBrain <- changeImageColoring(overlay = muBrain, color = "magenta", 
                               dye = "Cy5")
showImage(recolor(muBrain))

}
