% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinateGeneration.R
\name{scaleCoords}
\alias{scaleCoords}
\title{Scale coordinates to the size of the image}
\usage{
scaleCoords(object)
}
\arguments{
\item{object}{SpatialOverlay object}
}
\value{
SpatialOverlay object
}
\description{
Scale coordinates to the size of the image
}
\examples{

}
